Organize folder for fragment
